+++ ruby/ruby/Ice.rb
@@ -233,7 +233,8 @@ end
 #
 
 require_relative 'Ice/BuiltinSequences.rb'
-require_relative 'Ice/Connection.rb'
+require_relative 'Ice/CompressBatch.rb'
+require_relative 'Ice/ConnectionClose.rb'
 require_relative 'Ice/Context.rb'
 require_relative 'Ice/EndpointSelectionType.rb'
 require_relative 'Ice/EndpointTypes.rb'
